// src/phaser/types/SidTypes.ts
export enum SidName {
  DEDKI = 'DEDKI DEP',
  VERDO = 'VERDO DEP',
  KEPTA = 'KEPTA DEP',
  ANCOL = 'ANCOL DEP',
  BETES = 'BETES DEP',
  MIXUT = 'MIXUT DEP',
  RIGUS = 'RIGUS DEP',
  GOPUP = 'GOPUP DEP',
  SEDOG = 'SEDOG DEP',
  AVSEP = 'AVSEP DEP',
  IKLEN = 'IKLEN DEP',
  KISEP = 'KISEP DEP',
  URSAL = 'URSAL DEP',
  BOMET = 'BOMET DEP',
  MAVAN = 'MAVAN DEP',
  TEVAD = 'TEVAD DEP',
  DUSOM = 'DUSOM DEP',
  OAKVL = 'OAKVL DEP',
  PEMBA = 'PEMBA DEP',
  PERLO = 'PERLO DEP',
  TULEK = 'TULEK DEP',
  LAKES = 'LAKES DEP',
  EBKIN = 'EBKIN DEP',
  MATES = 'MATES DEP',
  NOSIK = 'NOSIK DEP',
  NUGOP = 'NUGOP DEP',
}